requirements.txt
pandas
numpy
scikit-learn
scipy
openpyxl
joblib
matplotlib
umap-learn
sentence-transformers  
plotly
pip
